requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
frozenlist==1.6.0
h11==0.16.0
idna==3.10
magic-filter==1.0.12
multidict==6.4.4
pillow==10.4.0
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
pytube==15.0.0
requests==2.32.3
sqlparse==0.5.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
